modificacion de resources

+++ database/migrations/2023_09_07_011549_create_mascota_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->string('nombre');
             $table->string('raza');
-            $table->integer('estatura cm');
-            $table->integer('peso');
-            $table->float('edad');
+            $table->float('estatura cm');
+            $table->float('peso');
+            $table->integer('edad');
             $table->timestamps();
         });
     }
